Fixing a simulation bug in line with published correction
In the original version of the paper, Figure 6 was based on data
which was generated by a script which had not taken into account
a bug in the underlying simulation. While the simulation bug was
fixed well before publication, the script for Figure 6 had not
been updated to account for the bugfix and the data had not been
regenerated. This commit updates the script accordingly and includes
an updated figure.